view cache documentation and code org changes
